Output from the new stack_builder.sh
Though the entire script has not been ran as you can see, this is a good place to start in debugging it. Besides, if you take a look, this stack building script was clearly dead on arrival. Will be exiting script after every milestone (denotated by five #'s in the script that resulted in this output) and manually (copy/pasting) a functional, though incomplete, script as a new commit here. All the work between commits/milestones will be studied and fixed locally before committing.